Node and eliminated some unnecessary imports
