+++ backend/src/node/router.py
@@ -1,0 +1,2 @@
+# Re-expose router object.
+from node.service import router
