+++ backend/src/timestamp/repository.py
@@ -3,9 +3,9 @@ from psycopg.connection_async import AsyncConnection
 from dataclasses import dataclass
 from dbutil import default_HTTP_exception
 from datetime import datetime
-from psycopg import sql, errors
+from psycopg import sql
 from psycopg.connection_async import AsyncConnection
-from psycopg.rows import class_row, scalar_row
+from psycopg.rows import scalar_row
 from psycopg import Error as PGError
 
 @dataclass
